# Optimise array::helpers::ForEach for small arrays.