@odlomax 2023-05-09
> Nice improvements! I like that you don't have the `arrayViewsCopy` anymore.

Thanks, Willem. Your suggestions make the code a lot cleaner (and hopefully readable). I think I know how forwarding works now!