@wdeconinck 2023-05-10
FYI, before you go deeper, I have already finished working on a field::for_each API.
I was just about to create a PR, but then you dropped this. I'll have to reintegrate.